Auto-Correction
The project was about getting the nearest correct word for the misspelled one. Here, we used levenshtein distance. First we computed the cost of replacing, deleting, inserting and switching characters then for each word in the list we check wether this word could be in our dictionary. In this project we did not care about the context of the sentence. For instance "I ate Pizza two vays ago". The nearest correct words for vays would be days/dye but here our model will not differentiate between both.